Scalable-Sentiment-Analysis-Pipeline-for-Customer-Feedback/requirements.txt
mysql-connector-python
nltk
pandas
